+++ packages/babel-preset-fbjs/plugins/inline-requires.js
@@ -1,156 +1,175 @@
 /**
- * Copyright (c) 2013-present, Facebook, Inc.
+ * Copyright (c) Facebook, Inc. and its affiliates.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
+ *
+ * @format
  */
 
 'use strict';
 
 /**
  * This transform inlines top-level require(...) aliases with to enable lazy
- * loading of dependencies.
+ * loading of dependencies. It is able to inline both single references and
+ * child property references.
  *
- * Continuing with the example above, this replaces all references to `Foo` in
- * the module to `require('ModuleFoo')`.
+ * For instance:
+ *     var Foo = require('foo');
+ *     f(Foo);
+ *
+ * Will be transformed into:
+ *     f(require('foo'));
+ *
+ * When the assigment expression has a property access, it will be inlined too,
+ * keeping the property. For instance:
+ *     var Bar = require('foo').bar;
+ *     g(Bar);
+ *
+ * Will be transformed into:
+ *     g(require('foo').bar);
+ *
+ * Destructuring also works the same way. For instance:
+ *     const {Baz} = require('foo');
+ *     h(Baz);
+ *
+ * Is also successfully inlined into:
+ *     g(require('foo').Baz);
  */
-module.exports = function fbjsInlineRequiresTransform(babel) {
-  var t = babel.types;
-
-  function buildRequireCall(name, inlineRequiredDependencyMap) {
-    var call = t.callExpression(
-      t.identifier('require'),
-      [t.stringLiteral(inlineRequiredDependencyMap.get(name))]
-    );
-    call.new = true;
-    return call;
-  }
-
-  function inlineRequire(path, inlineRequiredDependencyMap, identifierToPathsMap) {
-    var node = path.node;
-
-    if (path.isReferenced()) {
-      path.replaceWith(buildRequireCall(node.name, inlineRequiredDependencyMap));
-      var paths = identifierToPathsMap.get(node.name);
-      if (paths) {
-        paths.delete(path);
-      }
-    }
-  }
-
+module.exports = babel => {
   return {
+    name: 'inline-requires',
     visitor: {
-      Program(_, state) {
-        /**
-         * Map of require(...) aliases to module names.
-         *
-         * `Foo` is an alias for `require('ModuleFoo')` in the following example:
-         *   var Foo = require('ModuleFoo');
-         */
-        state.inlineRequiredDependencyMap = new Map();
+      Program: {
+        exit(path, state) {
+          var ignoredRequires = {};
+          var inlineableCalls = {require: true};
 
-        /**
-         * Map of variable names that have not yet been inlined.
-         * We track them in case we later remove their require()s,
-         * In which case we have to come back and update them.
-         */
-        state.identifierToPathsMap = new Map();
-      },
-
-      /**
-       * Collect top-level require(...) aliases.
-       */
-      CallExpression: function(path, state) {
-        var node = path.node;
-
-        if (isTopLevelRequireAlias(path)) {
-          var varName = path.parent.id.name;
-          var moduleName = node.arguments[0].value;
-          var inlineRequiredDependencyMap = state.inlineRequiredDependencyMap;
-          var identifierToPathsMap = state.identifierToPathsMap;
-
-          inlineRequiredDependencyMap.set(varName, moduleName);
-
-          // If we removed require() statements for variables we've already seen,
-          // We need to do a second pass on this program to replace them with require().
-          var maybePaths = identifierToPathsMap.get(varName);
-          if (maybePaths) {
-            maybePaths.forEach(path =>
-              inlineRequire(path, inlineRequiredDependencyMap, identifierToPathsMap));
-            maybePaths.delete(varName);
+          if (state.opts) {
+            if (state.opts.ignoredRequires) {
+              state.opts.ignoredRequires.forEach(function(name) {
+                ignoredRequires[name] = true;
+              });
+            }
+            if (state.opts.inlineableCalls) {
+              state.opts.inlineableCalls.forEach(function(name) {
+                inlineableCalls[name] = true;
+              });
+            }
           }
 
-          // Remove the declaration.
-          path.parentPath.parentPath.remove();
-          // And the associated binding in the scope.
-          path.scope.removeBinding(varName);
-        }
-      },
-
-      /**
-       * Inline require(...) aliases.
-       */
-      Identifier: function(path, state) {
-        var node = path.node;
-        var parent = path.parent;
-        var scope = path.scope;
-        var identifierToPathsMap = state.identifierToPathsMap;
-
-        if (!shouldInlineRequire(node, scope, state.inlineRequiredDependencyMap)) {
-          // Monitor this name in case we later remove its require().
-          // This won't happen often but if it does we need to come back and update here.
-          var paths = identifierToPathsMap.get(node.name);
-          if (paths) {
-            paths.add(path);
-          } else {
-            identifierToPathsMap.set(node.name, new Set([path]));
-          }
-
-          return;
-        }
-
-        if (
-          parent.type === 'AssignmentExpression' &&
-          path.isBindingIdentifier() &&
-          !scope.bindingIdentifierEquals(node.name, node)
-        ) {
-          throw new Error(
-            'Cannot assign to a require(...) alias, ' + node.name +
-            '. Line: ' + node.loc.start.line + '.'
+          path.scope.crawl();
+          path.traverse(
+            {CallExpression: call.bind(null, babel)},
+            {
+              ignoredRequires: ignoredRequires,
+              inlineableCalls: inlineableCalls,
+            }
           );
-        }
-
-        inlineRequire(path, state.inlineRequiredDependencyMap, identifierToPathsMap);
+        },
       },
     },
   };
 };
 
-function isTopLevelRequireAlias(path) {
-  return (
-    isRequireCall(path.node) &&
+function call(babel, path, state) {
+  var declaratorPath =
+    inlineableAlias(path, state) || inlineableMemberAlias(path, state);
+  var declarator = declaratorPath && declaratorPath.node;
+
+  if (declarator) {
+    var init = declarator.init;
+    var name = declarator.id && declarator.id.name;
+
+    var binding = declaratorPath.scope.getBinding(name);
+    var constantViolations = binding.constantViolations;
+    var thrown = false;
+
+    if (!constantViolations.length) {
+      deleteLocation(init);
+
+      babel.traverse(init, {
+        noScope: true,
+        enter: path => deleteLocation(path.node),
+      });
+
+      binding.referencePaths.forEach(ref => {
+        try {
+          ref.replaceWith(init);
+        } catch (err) {
+          thrown = true;
+        }
+      });
+
+      // If an error was thrown, it's most likely due to an invalid replacement
+      // happening (e.g. trying to replace a type annotation). It would usually
+      // be OK to ignore it, but to be safe, we will avoid removing the initial
+      // require.
+      if (!thrown) {
+        declaratorPath.remove();
+      }
+    }
+  }
+}
+
+function deleteLocation(node) {
+  delete node.start;
+  delete node.end;
+  delete node.loc;
+}
+
+function inlineableAlias(path, state) {
+  const isValid =
+    isInlineableCall(path.node, state) &&
     path.parent.type === 'VariableDeclarator' &&
     path.parent.id.type === 'Identifier' &&
     path.parentPath.parent.type === 'VariableDeclaration' &&
-    path.parentPath.parent.declarations.length === 1 &&
-    path.parentPath.parentPath.parent.type === 'Program'
-  );
+    path.parentPath.parentPath.parent.type === 'Program';
+
+  return isValid ? path.parentPath : null;
 }
 
-function shouldInlineRequire(node, scope, inlineRequiredDependencyMap) {
-  return (
-    inlineRequiredDependencyMap.has(node.name) &&
-    !scope.hasBinding(node.name, true /* noGlobals */)
-  );
+function inlineableMemberAlias(path, state) {
+  const isValid =
+    isInlineableCall(path.node, state) &&
+    path.parent.type === 'MemberExpression' &&
+    path.parentPath.parent.type === 'VariableDeclarator' &&
+    path.parentPath.parent.id.type === 'Identifier' &&
+    path.parentPath.parentPath.parent.type === 'VariableDeclaration' &&
+    path.parentPath.parentPath.parentPath.parent.type === 'Program';
+
+  return isValid ? path.parentPath.parentPath : null;
 }
 
-function isRequireCall(node) {
-  return (
-    !node.new &&
+function isInlineableCall(node, state) {
+  const isInlineable =
     node.type === 'CallExpression' &&
     node.callee.type === 'Identifier' &&
-    node.callee.name === 'require' &&
-    node['arguments'].length === 1 &&
-    node['arguments'][0].type === 'StringLiteral'
-  );
+    state.inlineableCalls.hasOwnProperty(node.callee.name) &&
+    node['arguments'].length >= 1;
+
+  // require('foo');
+  const isStandardCall =
+    isInlineable &&
+    node['arguments'][0].type === 'StringLiteral' &&
+    !state.ignoredRequires.hasOwnProperty(node['arguments'][0].value);
+
+  // require(require.resolve('foo'));
+  const isRequireResolveCall =
+    isInlineable &&
+    node['arguments'][0].type === 'CallExpression' &&
+    node['arguments'][0].callee.type === 'MemberExpression' &&
+    node['arguments'][0].callee.object.type === 'Identifier' &&
+    state.inlineableCalls.hasOwnProperty(
+      node['arguments'][0].callee.object.name
+    ) &&
+    node['arguments'][0].callee.property.type === 'Identifier' &&
+    node['arguments'][0].callee.property.name === 'resolve' &&
+    node['arguments'][0]['arguments'].length >= 1 &&
+    node['arguments'][0]['arguments'][0].type === 'StringLiteral' &&
+    !state.ignoredRequires.hasOwnProperty(
+      node['arguments'][0]['arguments'][0].value
+    );
+
+  return isStandardCall || isRequireResolveCall;
 }