+++ packages/eslint-config-fbjs/scripts/check-rules.js
@@ -1,4 +1,11 @@
 /**
+ * Copyright (c) 2013-present, Facebook, Inc.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+/**
  * Finds rules that are supported by ESLint but not defined in our config.
  */