+++ packages/fbjs-eslint-utils/index.js
@@ -1,1 +1,8 @@
+/**
+ * Copyright (c) 2013-present, Facebook, Inc.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
 throw new Error('This does not do anything. Require other modules.');