+++ packages/fbjs-scripts/gulp/module-map.js
@@ -7,7 +7,7 @@
 
 'use strict';
 
-var gutil = require('gulp-util');
+var PluginError = require('plugin-error');
 var through = require('through2');
 var fs = require('fs');
 var path = require('path');
@@ -19,7 +19,7 @@ var PLUGIN_NAME = 'module-map';
 module.exports = function(opts) {
   // Assume file is a string for now
   if (!opts || !('moduleMapFile' in opts && 'prefix' in opts)) {
-    throw new gutil.PluginError(
+    throw new PluginError(
       PLUGIN_NAME,
       'Missing options. Ensure you pass an object with `moduleMapFile` and `prefix`'
     );
@@ -35,7 +35,7 @@ module.exports = function(opts) {
     }
 
     if (file.isStream()) {
-      cb(new gutil.PluginError('module-map', 'Streaming not supported'));
+      cb(new PluginError('module-map', 'Streaming not supported'));
       return;
     }
 
@@ -46,7 +46,7 @@ module.exports = function(opts) {
       if (moduleMap.hasOwnProperty(name)) {
         this.emit(
           'error',
-          new gutil.PluginError(
+          new PluginError(
             PLUGIN_NAME,
             'Duplicate module found: ' + name + ' at ' + file.path + ' and ' +
               moduleMap[name]