+++ packages/fbjs-scripts/jest/createCacheKeyFunction.js
@@ -9,23 +9,45 @@
 
 const crypto = require('crypto');
 const fs = require('fs');
+const path = require('path');
 
-function buildCacheKey(files, base) {
-  return files.reduce(
-    (src, fileName) => src + fs.readFileSync(fileName),
-    base
-  );
+function getGlobalCacheKey(files, values) {
+  const presetVersion = require('../package').dependencies['babel-preset-fbjs'];
+
+  const chunks = [
+    process.env.NODE_ENV,
+    process.env.BABEL_ENV,
+    presetVersion,
+    ...values,
+    ...files.map(file => fs.readFileSync(file)),
+  ];
+
+  return chunks
+    .reduce(
+      (hash, chunk) => hash.update('\0', 'utf-8').update(chunk || ''),
+      crypto.createHash('md5')
+    )
+    .digest('hex');
 }
 
-module.exports = files => {
-  const presetVersion = require('../package').dependencies['babel-preset-fbjs'];
-  const cacheKey = buildCacheKey(files, presetVersion);
+function getCacheKeyFunction(globalCacheKey) {
   return (src, file, configString, options) => {
+    const {instrument, config} = options;
+    const rootDir = config && config.rootDir;
+
     return crypto
       .createHash('md5')
-      .update(cacheKey)
-      .update(src + file + configString)
-      .update(options && options.instrument ? 'instrument' : '')
+      .update(globalCacheKey)
+      .update('\0', 'utf8')
+      .update(src)
+      .update('\0', 'utf8')
+      .update(rootDir ? path.relative(config.rootDir, file) : '')
+      .update('\0', 'utf8')
+      .update(instrument ? 'instrument' : '')
       .digest('hex');
   };
+}
+
+module.exports = (files = [], values = []) => {
+  return getCacheKeyFunction(getGlobalCacheKey(files, values));
 };