+++ packages/signedsource/index.js
@@ -1,9 +1,12 @@
 /**
- * Copyright (c) 2015-present, Facebook, Inc.
+ * Copyright (c) Facebook, Inc. and its affiliates.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
+ *
+ * @format
  */
+
 'use strict';
 
 const crypto = require('crypto');
@@ -51,7 +54,7 @@ const SignedSource = {
    * Checks whether a file is signed *without* verifying the signature.
    */
   isSigned(data) {
-    return !PATTERN.exec(data);
+    return PATTERN.exec(data) != null;
   },
 
   /**
@@ -79,7 +82,7 @@ const SignedSource = {
     if (!matches) {
       throw new Error(
         'SignedSource.verifySignature(...): Cannot verify signature of an ' +
-        'unsigned file.'
+          'unsigned file.'
       );
     }
     const actual = matches[1];